Custom gates

BraketSimulator.jl defines some custom gates to extend what Braket.jl provides. These include the OpenQASM3 built-in gates gphase (as MultiQubitPhaseShift) and U (the single qubit three angle unitary).

BraketSimulator.U — Type

U(θ, ϕ, λ)

3-angle single qubit unitary gate. Equivalent to the OpenQASM3 built-in U gate.

BraketSimulator.MultiQubitPhaseShift — Type

MultiQubitPhaseShift{N}(angle)

Global phase shift on N qubits. Equivalent to the OpenQASM3 built-in gphase gate. Controls/negative controls applied to this gate control which states are rotated, so that Control(g::MultiQubitPhaseShift{2}) will apply the rotation to the |11> state.

BraketSimulator.MultiRZ — Type

MultiRz(angle)

Multi-qubit Z-rotation gate. The 2-qubit version is equivalent to the ZZ gate, and the single-qubit version is equivalent to the Rz gate.