Gates

BraketSimulators.jl provides many pre-implemented gates which can be used to build up circuits. For gates with angle parameters, you can supply Irrationals like π as arguments.

BraketSimulator.Gate — Type

Gate <: QuantumOperator

Abstract type representing a quantum gate.

BraketSimulator.AngledGate — Type

AngledGate{NA} <: Gate

Parametric type representing a quantum gate with NA angle parameters.

BraketSimulator.I — Type

BraketSimulator.I <: Gate
BraketSimulator.I() -> BraketSimulator.I

BraketSimulator.I gate.

BraketSimulator.X — Type

BraketSimulator.X <: Gate
BraketSimulator.X() -> BraketSimulator.X

BraketSimulator.X gate.

BraketSimulator.Y — Type

BraketSimulator.Y <: Gate
BraketSimulator.Y() -> BraketSimulator.Y

BraketSimulator.Y gate.

BraketSimulator.Z — Type

BraketSimulator.Z <: Gate
BraketSimulator.Z() -> BraketSimulator.Z

BraketSimulator.Z gate.

BraketSimulator.H — Type

BraketSimulator.H <: Gate
BraketSimulator.H() -> BraketSimulator.H

BraketSimulator.H gate.

BraketSimulator.Rx — Type

BraketSimulator.Rx <: AngledGate{1}
BraketSimulator.Rx(angles) -> BraketSimulator.Rx

BraketSimulator.Rx gate.

BraketSimulator.Ry — Type

BraketSimulator.Ry <: AngledGate{1}
BraketSimulator.Ry(angles) -> BraketSimulator.Ry

BraketSimulator.Ry gate.

BraketSimulator.Rz — Type

BraketSimulator.Rz <: AngledGate{1}
BraketSimulator.Rz(angles) -> BraketSimulator.Rz

BraketSimulator.Rz gate.

BraketSimulator.V — Type

BraketSimulator.V <: Gate
BraketSimulator.V() -> BraketSimulator.V

BraketSimulator.V gate.

BraketSimulator.Vi — Type

BraketSimulator.Vi <: Gate
BraketSimulator.Vi() -> BraketSimulator.Vi

BraketSimulator.Vi gate.

BraketSimulator.T — Type

BraketSimulator.T <: Gate
BraketSimulator.T() -> BraketSimulator.T

BraketSimulator.T gate.

BraketSimulator.Ti — Type

BraketSimulator.Ti <: Gate
BraketSimulator.Ti() -> BraketSimulator.Ti

BraketSimulator.Ti gate.

BraketSimulator.S — Type

BraketSimulator.S <: Gate
BraketSimulator.S() -> BraketSimulator.S

BraketSimulator.S gate.

BraketSimulator.Si — Type

BraketSimulator.Si <: Gate
BraketSimulator.Si() -> BraketSimulator.Si

BraketSimulator.Si gate.

BraketSimulator.U — Type

BraketSimulator.U <: AngledGate{3}
BraketSimulator.U(angles) -> BraketSimulator.U

BraketSimulator.U gate.

BraketSimulator.Unitary — Type

Unitary <: Gate
Unitary(matrix::Matrix{ComplexF64}) -> Unitary

Arbitrary unitary gate.

BraketSimulator.PhaseShift — Type

BraketSimulator.PhaseShift <: AngledGate{1}
BraketSimulator.PhaseShift(angles) -> BraketSimulator.PhaseShift

BraketSimulator.PhaseShift gate.

BraketSimulator.GPhase — Type

GPhase{N}(angle)

Global phase shift on N qubits. Equivalent to the OpenQASM3 built-in gphase gate. Controls/negative controls applied to this gate control which states are rotated, so that Control(g::GPhase{2}) will apply the rotation to the |11> state.

BraketSimulator.PRx — Type

BraketSimulator.PRx <: AngledGate{2}
BraketSimulator.PRx(angles) -> BraketSimulator.PRx

BraketSimulator.PRx gate.

BraketSimulator.GPi — Type

BraketSimulator.GPi <: AngledGate{1}
BraketSimulator.GPi(angles) -> BraketSimulator.GPi

BraketSimulator.GPi gate.

BraketSimulator.GPi2 — Type

BraketSimulator.GPi2 <: AngledGate{1}
BraketSimulator.GPi2(angles) -> BraketSimulator.GPi2

BraketSimulator.GPi2 gate.

BraketSimulator.XX — Type

BraketSimulator.XX <: AngledGate{1}
BraketSimulator.XX(angles) -> BraketSimulator.XX

BraketSimulator.XX gate.

BraketSimulator.XY — Type

BraketSimulator.XY <: AngledGate{1}
BraketSimulator.XY(angles) -> BraketSimulator.XY

BraketSimulator.XY gate.

BraketSimulator.YY — Type

BraketSimulator.YY <: AngledGate{1}
BraketSimulator.YY(angles) -> BraketSimulator.YY

BraketSimulator.YY gate.

BraketSimulator.ZZ — Type

BraketSimulator.ZZ <: AngledGate{1}
BraketSimulator.ZZ(angles) -> BraketSimulator.ZZ

BraketSimulator.ZZ gate.

BraketSimulator.ECR — Type

BraketSimulator.ECR <: Gate
BraketSimulator.ECR() -> BraketSimulator.ECR

BraketSimulator.ECR gate.

BraketSimulator.MS — Type

BraketSimulator.MS <: AngledGate{3}
BraketSimulator.MS(angles) -> BraketSimulator.MS

BraketSimulator.MS gate.

BraketSimulator.CPhaseShift — Type

BraketSimulator.CPhaseShift <: AngledGate{1}
BraketSimulator.CPhaseShift(angles) -> BraketSimulator.CPhaseShift

BraketSimulator.CPhaseShift gate.

BraketSimulator.CPhaseShift00 — Type

BraketSimulator.CPhaseShift00 <: AngledGate{1}
BraketSimulator.CPhaseShift00(angles) -> BraketSimulator.CPhaseShift00

BraketSimulator.CPhaseShift00 gate.

BraketSimulator.CPhaseShift01 — Type

BraketSimulator.CPhaseShift01 <: AngledGate{1}
BraketSimulator.CPhaseShift01(angles) -> BraketSimulator.CPhaseShift01

BraketSimulator.CPhaseShift01 gate.

BraketSimulator.CPhaseShift10 — Type

BraketSimulator.CPhaseShift10 <: AngledGate{1}
BraketSimulator.CPhaseShift10(angles) -> BraketSimulator.CPhaseShift10

BraketSimulator.CPhaseShift10 gate.

BraketSimulator.CNot — Type

BraketSimulator.CNot <: Gate
BraketSimulator.CNot() -> BraketSimulator.CNot

BraketSimulator.CNot gate.

BraketSimulator.CY — Type

BraketSimulator.CY <: Gate
BraketSimulator.CY() -> BraketSimulator.CY

BraketSimulator.CY gate.

BraketSimulator.CZ — Type

BraketSimulator.CZ <: Gate
BraketSimulator.CZ() -> BraketSimulator.CZ

BraketSimulator.CZ gate.

BraketSimulator.CV — Type

BraketSimulator.CV <: Gate
BraketSimulator.CV() -> BraketSimulator.CV

BraketSimulator.CV gate.

BraketSimulator.Swap — Type

BraketSimulator.Swap <: Gate
BraketSimulator.Swap() -> BraketSimulator.Swap

BraketSimulator.Swap gate.

BraketSimulator.PSwap — Type

BraketSimulator.PSwap <: AngledGate{1}
BraketSimulator.PSwap(angles) -> BraketSimulator.PSwap

BraketSimulator.PSwap gate.

BraketSimulator.ISwap — Type

BraketSimulator.ISwap <: Gate
BraketSimulator.ISwap() -> BraketSimulator.ISwap

BraketSimulator.ISwap gate.

BraketSimulator.CCNot — Type

BraketSimulator.CCNot <: Gate
BraketSimulator.CCNot() -> BraketSimulator.CCNot

BraketSimulator.CCNot gate.

BraketSimulator.CSwap — Type

BraketSimulator.CSwap <: Gate
BraketSimulator.CSwap() -> BraketSimulator.CSwap

BraketSimulator.CSwap gate.