Noises

BraketSimulators.jl provides many pre-implemented noise channels which can be applied to circuits. Noisy circuits can be simulated used the density matrix simulator.

BraketSimulator.Noise — Type

Noise <: QuantumOperator

Abstract type representing a quantum noise operation.

BraketSimulator.Kraus — Type

Kraus <: Noise

Kraus noise operation.

BraketSimulator.BitFlip — Type

BitFlip <: Noise

BitFlip noise operation.

BraketSimulator.PhaseFlip — Type

PhaseFlip <: Noise

PhaseFlip noise operation.

BraketSimulator.PauliChannel — Type

PauliChannel <: Noise

PauliChannel noise operation.

BraketSimulator.TwoQubitPauliChannel — Type

TwoQubitPauliChannel <: Noise

Pauli channel noise operation on two qubits.

BraketSimulator.MultiQubitPauliChannel — Type

MultiQubitPauliChannel{N} <: Noise

Pauli channel noise operation on N qubits.

BraketSimulator.Depolarizing — Type

Depolarizing <: Noise

Depolarizing noise operation.

BraketSimulator.PhaseDamping — Type

PhaseDamping <: Noise

PhaseDamping noise operation.

BraketSimulator.AmplitudeDamping — Type

AmplitudeDamping <: Noise

AmplitudeDamping noise operation.

BraketSimulator.GeneralizedAmplitudeDamping — Type

GeneralizedAmplitudeDamping <: Noise

GeneralizedAmplitudeDamping noise operation.

BraketSimulator.TwoQubitDepolarizing — Type

TwoQubitDepolarizing <: Noise

TwoQubitDepolarizing noise operation.

BraketSimulator.TwoQubitDephasing — Type

TwoQubitDephasing <: Noise

TwoQubitDephasing noise operation.