Results

BraketSimulator.Result — Type

Result

Abstract type representing a measurement to perform on a Circuit.

See also: Expectation, Variance, Sample, Probability, DensityMatrix, and Amplitude.

BraketSimulator.AdjointGradient — Type

AdjointGradient <: Result

Struct which represents a gradient computation using the adjoint differentiation method on a Circuit.

BraketSimulator.Expectation — Type

BraketSimulator.Expectation <: Result

Struct which represents a expectation measurement on a Circuit.

BraketSimulator.Variance — Type

BraketSimulator.Variance <: Result

Struct which represents a variance measurement on a Circuit.

BraketSimulator.Sample — Type

BraketSimulator.Sample <: Result

Struct which represents a sample measurement on a Circuit.

BraketSimulator.Amplitude — Type

Amplitude <: Result

Struct which represents an amplitude measurement on a Circuit.

BraketSimulator.Probability — Type

BraketSimulator.Probability <: Result

Struct which represents a probability measurement on a Circuit.

BraketSimulator.DensityMatrix — Type

BraketSimulator.DensityMatrix <: Result

Struct which represents a densitymatrix measurement on a Circuit.

BraketSimulator.StateVector — Type

StateVector <: Result

Struct which represents a state vector measurement on a Circuit.